+++ CHANGELOG.md
@@ -2,11 +2,13 @@
 
 ## 0.6.1 Under development
 
+- Enh #27: Add `positiveLike()` method in `BaseValidator` for validating positive-like numbers with optional maximum constraint, along with related tests (@terabytesoftw)
+- Enh #28: Enhance `positiveLike()` method to include minimum value validation and update related tests (@terabytesoftw)
+
 ## 0.6.0 December 27, 2025
 
 - Enh #25: Add `normalizeAttributes()` method in `BaseAttributes` helper for normalization of attribute keys and values with support for encoding and JSON serialization, along with comprehensive tests (@terabytesoftw)
 - Bug #26: Rename exception test methods to clarify expected exceptions (@terabytesoftw)
-- Enh #27: Add `positiveLike()` method in `BaseValidator` for validating positive-like numbers with optional maximum constraint, along with related tests (@terabytesoftw)
 
 ## 0.5.3 December 26, 2025