+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # ChangeLog
 
-## 0.6.1 Under development
+## 0.6.1 December 31, 2025
 
 - Enh #27: Add `positiveLike()` method in `BaseValidator` for validating positive-like numbers with optional maximum constraint, along with related tests (@terabytesoftw)
 - Enh #28: Enhance `positiveLike()` method to include minimum value validation and update related tests (@terabytesoftw)